requirements.txt
python-dotenv
pandas
numpy
scikit-learn
matplotlib
seaborn
pymongo
certifi
pymongo[srv]==3.6

# -e .